Update Framework
Originally, this project was done with Vue 2.x and Vue CLI, and made use of Babel. Along the way, several dependencies were deprecated, and the project would no lunger run, as it was.

I scaffolded out a new Vue 3 project with Vite, and now we're back on track, at least getting the basic layout to display again.