Update API Calls - MVP
OpenWeatherMap updated their "current weather" API endpoint, and it no longer accepts a loose variety of location names. Now, it only accepts a string containing the city, state (in some situations), and valid country code (and results may still be unpredictable) -- OR, you can give it exact latitude and longitude. Of course, nobody knows their latitude and longitude, so you have to send your location names to a geocoding API endpoint first, get the latitude and longitude, then send that back to the "current weather" endpoint.

This commit takes care of all that. Still need to add some better error handling and styling, but this is a mostly functional first-step.