Submit Query on Change
Instead of submitting the user's query to the OpenWeather API when they hit 'enter,' I'm having it happen on each change of the input's value (eg, as the user types). I've modified the function that handles that API call to include a "debounce" so we don't clobber the API too much.